refactor: restructure routing and layout components for better organization and maintainability
